# satelite-catalogo-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CC:
    properties:
      id:
        type: string
      jefe:
        type: string
      nombre:
        type: string
      numero:
        type: integer
    type: object
...
  models.Product:
    properties:
      UM:
        type: string
      categoria:
        type: string
      convenio_marco:
        type: string
      descripcion:
        type: string
      fecha_actualizacion:
        type: string
      id:
        type: string
      id_categoria:
        type: string
      id_convenio:
        type: string
      id_product:
        type: string
      licitacion:
        type: string
      marca:
        type: string
      modelo:
        type: string
      nombre_proveedor:
        type: string
      number:
        type: string
      precio:
        type: number
      region:
        type: string
      rut_proveedor:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Catalogo service.
  title: Catalogo API
  version: "1.0"
paths:
  /archivos/{filepath}:
    get:
      description: Serves files from the uploads directory in a safe manner
      parameters:
      - description: File path
        in: path
        name: filepath
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Serve uploaded file
      tags:
      - files
  /cc/:
    get:
      description: Returns all Centros de Costo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CC'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List centros de costo
      tags:
      - cc
    post:
      consumes:
      - application/json
      description: Creates a new Centro de Costo
      parameters:
      - description: Centro de Costo info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CC'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create centro de costo
      tags:
      - cc
  /cc/{id}:
    delete:
      description: Deletes a Centro de Costo by ID
      parameters:
      - description: Centro de Costo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete centro de costo
      tags:
      - cc
    get:
      consumes:
      - application/json
      description: Returns a Centro de Costo by its ID
      parameters:
      - description: Centro de Costo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CC'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get centro de costo by ID
      tags:
      - cc
    put:
      consumes:
      - application/json
      description: Updates a Centro de Costo by ID
      parameters:
      - description: Centro de Costo ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Update centro de costo
      tags:
      - cc
  /product/:
    get:
      description: Returns all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product with the provided information
      parameters:
      - description: Product info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create product
      tags:
      - products
  /product/paginated:
    get:
      description: Returns paginated products
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List products paginated
      tags:
      - products
...
swagger: "2.0"